MailboxManagerImp fixed ZAL compatibility for all Zimbra versions.
